requirements.txt
Pillow==4.3.0
aiohttp==2.2.5
async_timeout==2.0.0
asyncpg==0.12.0
beautifulsoup4==4.6.0
python_dateutil==2.6.1
Logbook==1.1.0
